+++ timeline/models.py
@@ -29,8 +29,8 @@ DEFAULT_MESSAGES = {
 OBJ_REF_UNNAMED = '(unnamed)'
 
 
-class ProjectEventManager(models.Manager):
-    """Manager for custom table-level ProjectEvent queries"""
+class TimelineEventManager(models.Manager):
+    """Manager for custom table-level TimelineEvent queries"""
 
     def get_object_events(
         self, project, object_model, object_uuid, order_by='-pk'
@@ -44,7 +44,7 @@ class ProjectEventManager(models.Manager):
         :param order_by: Ordering (default = pk descending)
         :return: QuerySet
         """
-        return ProjectEvent.objects.filter(
+        return TimelineEvent.objects.filter(
             project=project,
             event_objects__object_model=object_model,
             event_objects__object_uuid=object_uuid,
@@ -56,7 +56,7 @@ class ProjectEventManager(models.Manager):
 
         :param search_terms: Search terms (list of strings)
         :param keywords: Optional search keywords as key/value pairs (dict)
-        :return: QuerySet of ProjectEvent objects
+        :return: QuerySet of TimelineEvent objects
         """
         search_limit = getattr(settings, 'TIMELINE_SEARCH_LIMIT', 250)
         term_query = Q()
@@ -73,7 +73,7 @@ class ProjectEventManager(models.Manager):
         return items[:search_limit]
 
 
-class ProjectEvent(models.Model):
+class TimelineEvent(models.Model):
     """
     Class representing a Project event. Can also be a site-wide event not linked
     to a specific project.
@@ -137,7 +137,7 @@ class ProjectEvent(models.Model):
     )
 
     # Set manager for custom queries
-    objects = ProjectEventManager()
+    objects = TimelineEventManager()
 
     def __str__(self):
         return '{}{}{}'.format(
@@ -147,7 +147,7 @@ class ProjectEvent(models.Model):
         )
 
     def __repr__(self):
-        return 'ProjectEvent({})'.format(
+        return 'TimelineEvent({})'.format(
             ', '.join(repr(v) for v in self.get_repr_values())
         )
 
@@ -180,9 +180,9 @@ class ProjectEvent(models.Model):
         :param label: Label for the object in the event description (string)
         :param name: Name or title of the object (string)
         :param extra_data: Additional data related to object (dict, optional)
-        :return: ProjectEventObjectRef object
+        :return: TimelineEventObjectRef object
         """
-        ref = ProjectEventObjectRef()
+        ref = TimelineEventObjectRef()
         ref.event = self
         ref.label = label
         if not name:
@@ -207,7 +207,7 @@ class ProjectEvent(models.Model):
         :param status_type: Status type string (see EVENT_STATUS_TYPES)
         :param status_desc: Description string (optional)
         :param extra_data: Extra data for the status (dict, optional)
-        :return: ProjectEventStatus object
+        :return: TimelineEventStatus object
         :raise: TypeError if status_type is invalid
         """
         if status_type not in EVENT_STATUS_TYPES:
@@ -217,7 +217,7 @@ class ProjectEvent(models.Model):
                 )
             )
 
-        status = ProjectEventStatus()
+        status = TimelineEventStatus()
         status.event = self
         status.status_type = status_type
         status.description = (
@@ -229,13 +229,13 @@ class ProjectEvent(models.Model):
         return status
 
 
-class ProjectEventObjectRef(models.Model):
+class TimelineEventObjectRef(models.Model):
     """Class representing a reference to an object (existing or removed)
     related to a Timeline event status"""
 
     #: Event to which the object belongs
     event = models.ForeignKey(
-        ProjectEvent,
+        TimelineEvent,
         related_name='event_objects',
         help_text='Event to which the object belongs',
         on_delete=models.CASCADE,
@@ -283,17 +283,17 @@ class ProjectEventObjectRef(models.Model):
 
     def __repr__(self):
         values = self.event.get_repr_values() + [self.name]
-        return 'ProjectEventObjectRef({})'.format(
+        return 'TimelineEventObjectRef({})'.format(
             ', '.join(repr(v) for v in values)
         )
 
 
-class ProjectEventStatus(models.Model):
+class TimelineEventStatus(models.Model):
     """Class representing a Timeline event status"""
 
     #: Event to which the status change belongs
     event = models.ForeignKey(
-        ProjectEvent,
+        TimelineEvent,
         related_name='status_changes',
         help_text='Event to which the status change belongs',
         on_delete=models.CASCADE,
@@ -335,7 +335,7 @@ class ProjectEventStatus(models.Model):
 
     def __repr__(self):
         values = self.event.get_repr_values() + [self.status_type]
-        return 'ProjectEventStatus({})'.format(
+        return 'TimelineEventStatus({})'.format(
             ', '.join(repr(v) for v in values)
         )