+++ timeline/tests/test_models.py
@@ -13,9 +13,9 @@ from projectroles.tests.test_models import (
 )
 
 from timeline.models import (
-    ProjectEvent,
-    ProjectEventObjectRef,
-    ProjectEventStatus,
+    TimelineEvent,
+    TimelineEventObjectRef,
+    TimelineEventStatus,
     OBJ_REF_UNNAMED,
 )
 
@@ -29,8 +29,8 @@ PROJECT_TYPE_CATEGORY = SODAR_CONSTANTS['PROJECT_TYPE_CATEGORY']
 PROJECT_TYPE_PROJECT = SODAR_CONSTANTS['PROJECT_TYPE_PROJECT']
 
 
-class ProjectEventMixin:
-    """Helper mixin for ProjectEvent creation"""
+class TimelineEventMixin:
+    """Helper mixin for TimelineEvent creation"""
 
     @classmethod
     def make_event(
@@ -54,13 +54,13 @@ class ProjectEventMixin:
             'extra_data': extra_data or {},
             'plugin': plugin,
         }
-        result = ProjectEvent(**values)
+        result = TimelineEvent(**values)
         result.save()
         return result
 
 
-class ProjectEventObjectRefMixin:
-    """Helper mixin for ProjectEventObjectRef creation"""
+class TimelineEventObjectRefMixin:
+    """Helper mixin for TimelineEventObjectRef creation"""
 
     @classmethod
     def make_object_ref(cls, event, obj, label, name, uuid, extra_data=None):
@@ -72,13 +72,13 @@ class ProjectEventObjectRefMixin:
             'object_uuid': uuid,
             'extra_data': extra_data or {},
         }
-        result = ProjectEventObjectRef(**values)
+        result = TimelineEventObjectRef(**values)
         result.save()
         return result
 
 
-class ProjectEventStatusMixin:
-    """Helper mixin for ProjectEventStatus creation"""
+class TimelineEventStatusMixin:
+    """Helper mixin for TimelineEventStatus creation"""
 
     @classmethod
     def make_event_status(
@@ -90,12 +90,12 @@ class ProjectEventStatusMixin:
             'description': description,
             'extra_data': extra_data or {'test': 'test'},
         }
-        result = ProjectEventStatus(**values)
+        result = TimelineEventStatus(**values)
         result.save()
         return result
 
 
-class ProjectEventTestBase(
+class TimelineEventTestBase(
     ProjectMixin, RoleMixin, RoleAssignmentMixin, TestCase
 ):
     def setUp(self):
@@ -112,11 +112,11 @@ class ProjectEventTestBase(
         )
 
 
-class TestProjectEvent(
-    ProjectEventMixin,
-    ProjectEventStatusMixin,
-    ProjectEventObjectRefMixin,
-    ProjectEventTestBase,
+class TestTimelineEvent(
+    TimelineEventMixin,
+    TimelineEventStatusMixin,
+    TimelineEventObjectRefMixin,
+    TimelineEventTestBase,
 ):
     def setUp(self):
         super().setUp()
@@ -140,7 +140,7 @@ class TestProjectEvent(
         )
 
     def test_initialization(self):
-        """Test ProjectEvent initialization"""
+        """Test TimelineEvent initialization"""
         expected = {
             'id': self.event.pk,
             'project': self.project.pk,
@@ -156,7 +156,7 @@ class TestProjectEvent(
         self.assertEqual(model_to_dict(self.event), expected)
 
     def test_initialization_no_project(self):
-        """Test ProjectEvent initialization with no project"""
+        """Test TimelineEvent initialization with no project"""
         self.event = self.make_event(
             project=None,
             app='projectroles',
@@ -181,7 +181,7 @@ class TestProjectEvent(
         self.assertEqual(model_to_dict(self.event), expected)
 
     def test_initialization_no_user(self):
-        """Test ProjectEvent initialization with no user"""
+        """Test TimelineEvent initialization with no user"""
         self.event = self.make_event(
             project=self.project,
             app='projectroles',
@@ -206,7 +206,7 @@ class TestProjectEvent(
         self.assertEqual(model_to_dict(self.event), expected)
 
     def test_initialization_plugin(self):
-        """Test ProjectEvent initialization with specific plugin name"""
+        """Test TimelineEvent initialization with specific plugin name"""
         self.event = self.make_event(
             project=None,
             app='projectroles',
@@ -232,26 +232,26 @@ class TestProjectEvent(
         self.assertEqual(model_to_dict(self.event), expected)
 
     def test_str(self):
-        """Test ProjectEvent __str__()"""
+        """Test TimelineEvent __str__()"""
         expected = 'TestProject: test_event/owner'
         self.assertEqual(str(self.event), expected)
 
     def test_str_no_user(self):
-        """Test ProjectEvent __str__() with no user"""
+        """Test TimelineEvent __str__() with no user"""
         self.event.user = None
         self.event.save()
         expected = 'TestProject: test_event'
         self.assertEqual(str(self.event), expected)
 
     def test_str_no_project(self):
-        """Test ProjectEvent __str__() with no project"""
+        """Test TimelineEvent __str__() with no project"""
         self.event.project = None
         self.event.save()
         expected = 'test_event/owner'
         self.assertEqual(str(self.event), expected)
 
     def test_str_no_project_no_user(self):
-        """Test ProjectEvent __str__() with no project or user"""
+        """Test TimelineEvent __str__() with no project or user"""
         self.event.project = None
         self.event.user = None
         self.event.save()
@@ -259,30 +259,30 @@ class TestProjectEvent(
         self.assertEqual(str(self.event), expected)
 
     def test_repr(self):
-        """Test ProjectEventStatus __repr__()"""
-        expected = "ProjectEvent('TestProject', 'test_event', 'owner')"
+        """Test TimelineEvent __repr__()"""
+        expected = "TimelineEvent('TestProject', 'test_event', 'owner')"
         self.assertEqual(repr(self.event), expected)
 
     def test_repr_no_user(self):
-        """Test ProjectEventStatus __repr__() with no user"""
+        """Test TimelineEvent __repr__() with no user"""
         self.event.user = None
         self.event.save()
-        expected = "ProjectEvent('TestProject', 'test_event', 'N/A')"
+        expected = "TimelineEvent('TestProject', 'test_event', 'N/A')"
         self.assertEqual(repr(self.event), expected)
 
     def test_repr_no_project(self):
-        """Test ProjectEventStatus __repr__() with no project"""
+        """Test TimelineEvent __repr__() with no project"""
         self.event.project = None
         self.event.save()
-        expected = "ProjectEvent('N/A', 'test_event', 'owner')"
+        expected = "TimelineEvent('N/A', 'test_event', 'owner')"
         self.assertEqual(repr(self.event), expected)
 
     def test_repr_no_project_no_user(self):
-        """Test ProjectEventStatus __repr__() with no project or user"""
+        """Test TimelineEvent __repr__() with no project or user"""
         self.event.project = None
         self.event.user = None
         self.event.save()
-        expected = "ProjectEvent('N/A', 'test_event', 'N/A')"
+        expected = "TimelineEvent('N/A', 'test_event', 'N/A')"
         self.assertEqual(repr(self.event), expected)
 
     def test_add_object(self):
@@ -316,31 +316,31 @@ class TestProjectEvent(
         self.assertEqual(new_obj.name, OBJ_REF_UNNAMED)
 
     def test_find_name(self):
-        """Test ProjectEvent.find() with event name"""
-        objects = ProjectEvent.objects.find(['test_event'])
+        """Test TimelineEvent.find() with event name"""
+        objects = TimelineEvent.objects.find(['test_event'])
         self.assertEqual(len(objects), 1)
         self.assertEqual(objects[0], self.event)
 
     def test_find_description(self):
-        """Test ProjectEvent.find() with event description"""
-        objects = ProjectEvent.objects.find(['description'])
+        """Test TimelineEvent.find() with event description"""
+        objects = TimelineEvent.objects.find(['description'])
         self.assertEqual(len(objects), 1)
         self.assertEqual(objects[0], self.event)
 
     def test_find_object(self):
-        """Test ProjectEvent.find() with object reference"""
-        objects = ProjectEvent.objects.find(['test_object_name'])
+        """Test TimelineEvent.find() with object reference"""
+        objects = TimelineEvent.objects.find(['test_object_name'])
         self.assertEqual(len(objects), 1)
         self.assertEqual(objects[0], self.event)
 
     def test_find_fail(self):
-        """Test ProjectEvent.find() with no matches"""
-        objects = ProjectEvent.objects.find(['asdfasdfafasdf'])
+        """Test TimelineEvent.find() with no matches"""
+        objects = TimelineEvent.objects.find(['asdfasdfafasdf'])
         self.assertEqual(len(objects), 0)
 
 
-class TestProjectEventObjectRef(
-    ProjectEventMixin, ProjectEventObjectRefMixin, ProjectEventTestBase
+class TestTimelineEventObjectRef(
+    TimelineEventMixin, TimelineEventObjectRefMixin, TimelineEventTestBase
 ):
     def setUp(self):
         super().setUp()
@@ -363,7 +363,7 @@ class TestProjectEventObjectRef(
         )
 
     def test_initialization(self):
-        """Test ProjectEventObject initialization"""
+        """Test TimelineEventObjectRef initialization"""
         expected = {
             'id': self.obj_ref.pk,
             'event': self.event.pk,
@@ -376,21 +376,21 @@ class TestProjectEventObjectRef(
         self.assertEqual(model_to_dict(self.obj_ref), expected)
 
     def test__str__(self):
-        """Test ProjectEventObject __str__()"""
+        """Test TimelineEventObjectRef __str__()"""
         expected = 'TestProject: test_event/owner (test_name)'
         self.assertEqual(str(self.obj_ref), expected)
 
     def test__repr__(self):
-        """Test ProjectEventObject __repr__()"""
+        """Test TimelineEventObjectRef __repr__()"""
         expected = (
-            "ProjectEventObjectRef('TestProject', 'test_event', "
+            "TimelineEventObjectRef('TestProject', 'test_event', "
             "'owner', 'test_name')"
         )
         self.assertEqual(repr(self.obj_ref), expected)
 
     def test_get_object_events(self):
-        """Test get_object_events() in ProjectEventManager"""
-        events = ProjectEvent.objects.get_object_events(
+        """Test get_object_events() in TimelineEventManager"""
+        events = TimelineEvent.objects.get_object_events(
             project=self.project,
             object_model=self.obj_ref.object_model,
             object_uuid=self.obj_ref.object_uuid,
@@ -399,8 +399,8 @@ class TestProjectEventObjectRef(
         self.assertEqual(events[0], self.event)
 
 
-class TestProjectEventStatus(
-    ProjectEventMixin, ProjectEventStatusMixin, ProjectEventTestBase
+class TestTimelineEventStatus(
+    TimelineEventMixin, TimelineEventStatusMixin, TimelineEventTestBase
 ):
     def setUp(self):
         super().setUp()
@@ -427,7 +427,7 @@ class TestProjectEventStatus(
         )
 
     def test_initialization(self):
-        """Test ProjectEventStatus init"""
+        """Test TimelineEventStatus init"""
         expected = {
             'id': self.event_status_ok.pk,
             'sodar_uuid': self.event_status_ok.sodar_uuid,
@@ -439,7 +439,7 @@ class TestProjectEventStatus(
         self.assertEqual(model_to_dict(self.event_status_ok), expected)
 
     def test_initialization_no_user(self):
-        """Test ProjectEventStatus without user"""
+        """Test TimelineEventStatus without user"""
         expected = {
             'id': self.event_status_ok.pk,
             'sodar_uuid': self.event_status_ok.sodar_uuid,
@@ -460,7 +460,7 @@ class TestProjectEventStatus(
         self.assertEqual(model_to_dict(self.event_status_ok), expected)
 
     def test__str__(self):
-        """Test ProjectEventStatus __str__()"""
+        """Test TimelineEventStatus __str__()"""
         expected = 'TestProject: test_event/owner (OK)'
         self.assertEqual(str(self.event_status_ok), expected)
 
@@ -472,9 +472,9 @@ class TestProjectEventStatus(
         self.assertEqual(str(self.event_status_ok), expected)
 
     def test__repr__(self):
-        """Test ProjectEventStatus __repr__()"""
+        """Test TimelineEventStatus __repr__()"""
         expected = (
-            "ProjectEventStatus('TestProject', 'test_event', 'owner', 'OK')"
+            "TimelineEventStatus('TestProject', 'test_event', 'owner', 'OK')"
         )
         self.assertEqual(repr(self.event_status_ok), expected)
 
@@ -483,12 +483,12 @@ class TestProjectEventStatus(
         self.event.user = None
         self.event.save()
         expected = (
-            "ProjectEventStatus('TestProject', 'test_event', 'N/A', 'OK')"
+            "TimelineEventStatus('TestProject', 'test_event', 'N/A', 'OK')"
         )
         self.assertEqual(repr(self.event_status_ok), expected)
 
     def test_get_status(self):
-        """Test the get_status() function of ProjectEvent"""
+        """Test TimelineEventStatus get_status()"""
         status = self.event.get_status()
         expected = {
             'id': status.pk,
@@ -501,25 +501,25 @@ class TestProjectEventStatus(
         self.assertEqual(model_to_dict(status), expected)
 
     def test_get_timestamp(self):
-        """Test the get_timestamp() function of ProjectEvent"""
+        """Test TimelineEventStatus get_timestamp()"""
         timestamp = self.event.get_timestamp()
         self.assertEqual(timestamp, self.event_status_ok.timestamp)
 
     def test_get_status_changes(self):
-        """Test the get_status_changes() function of ProjectEvent"""
+        """Test TimelineEventStatus get_status_changes()"""
         status_changes = self.event.get_status_changes()
         self.assertEqual(status_changes.count(), 2)
         self.assertEqual(status_changes[0], self.event_status_submit)
 
     def test_get_status_changes_reverse(self):
-        """Test the get_status_changes() function of ProjectEvent with
+        """Test the get_status_changes() function of TimelineEvent with
         reverse=True"""
         status_changes = self.event.get_status_changes(reverse=True)
         self.assertEqual(status_changes.count(), 2)
         self.assertEqual(status_changes[0], self.event_status_ok)
 
     def test_set_status(self):
-        """Test the set_status() function of ProjectEvent"""
+        """Test TimelineEventStatus set_status()"""
         new_status = self.event.set_status(
             'FAILED', status_desc='FAILED', extra_data={'test_key': 'test_val'}
         )