+++ timeline/views_ajax.py
@@ -15,7 +15,7 @@ from projectroles.views_ajax import (
     SODARBaseAjaxView,
 )
 
-from timeline.models import ProjectEvent, ProjectEventStatus
+from timeline.models import TimelineEvent, TimelineEventStatus
 from timeline.templatetags.timeline_tags import get_status_style
 
 
@@ -40,7 +40,7 @@ class EventDetailMixin:
         """
         Return event details.
 
-        :param event: ProjectEvent object
+        :param event: TimelineEvent object
         :param request: Django request object
         :return: Dict
         """
@@ -154,8 +154,8 @@ class EventExtraDataMixin:
     def get_event_extra(self, event, status=None):
         """
         Return event extra data.
-        :param event: ProjectEvent object
-        :param status: ProjectEventStatus object
+        :param event: TimelineEvent object
+        :param status: TimelineEventStatus object
         :return: JSON-serializable dict
         """
         extra = status.extra_data if status else event.extra_data
@@ -178,8 +178,8 @@ class ProjectEventDetailAjaxView(EventDetailMixin, SODARBaseProjectAjaxView):
     permission_required = 'timeline.view_timeline'
 
     def get(self, request, *args, **kwargs):
-        event = ProjectEvent.objects.filter(
-            sodar_uuid=self.kwargs['projectevent']
+        event = TimelineEvent.objects.filter(
+            sodar_uuid=self.kwargs['timelineevent']
         ).first()
         if event.classified and not request.user.has_perm(
             'timeline.view_classified_event', event.project
@@ -194,8 +194,8 @@ class ProjectEventExtraAjaxView(EventExtraDataMixin, SODARBaseProjectAjaxView):
     permission_required = 'timeline.view_event_extra_data'
 
     def get(self, request, *args, **kwargs):
-        event = ProjectEvent.objects.filter(
-            sodar_uuid=self.kwargs['projectevent']
+        event = TimelineEvent.objects.filter(
+            sodar_uuid=self.kwargs['timelineevent']
         ).first()
         if event.classified and not request.user.has_perm(
             'timeline.view_classified_event', event.project
@@ -211,8 +211,8 @@ class SiteEventDetailAjaxView(EventDetailMixin, SODARBasePermissionAjaxView):
     permission_required = 'timeline.view_site_timeline'
 
     def get(self, request, *args, **kwargs):
-        event = ProjectEvent.objects.filter(
-            sodar_uuid=self.kwargs['projectevent']
+        event = TimelineEvent.objects.filter(
+            sodar_uuid=self.kwargs['timelineevent']
         ).first()
         if event.classified and not request.user.has_perm(
             'timeline.view_classified_site_event'
@@ -227,8 +227,8 @@ class SiteEventExtraAjaxView(EventExtraDataMixin, SODARBasePermissionAjaxView):
     permission_required = 'timeline.view_event_extra_data'
 
     def get(self, request, *args, **kwargs):
-        event = ProjectEvent.objects.filter(
-            sodar_uuid=self.kwargs['projectevent']
+        event = TimelineEvent.objects.filter(
+            sodar_uuid=self.kwargs['timelineevent']
         ).first()
         if event.classified and not request.user.has_perm(
             'timeline.view_classified_site_event'
@@ -241,7 +241,7 @@ class EventStatusExtraAjaxView(EventExtraDataMixin, SODARBaseAjaxView):
     """Ajax view for retrieving event status extra data for events"""
 
     def get(self, request, *args, **kwargs):
-        status = ProjectEventStatus.objects.filter(
+        status = TimelineEventStatus.objects.filter(
             sodar_uuid=self.kwargs['eventstatus']
         ).first()
         event = status.event