+++ CHANGELOG.rst
@@ -60,6 +60,7 @@ Changed
     - Raise ``update_cache()`` exception for ``synccache`` in debug mode (#1375)
 - **Timeline**
     - Update ``get_object_link()`` usage for ``PluginObjectLink`` return data (#1343)
+    - Rename ``ProjectEvent*`` models to ``TimelineEvent*`` (#1414)
 - **Userprofile**
     - Disable global user settings on target site in ``UserSettingsForm`` (#1329)